+++ src/components/Containers/QRCode/index.jsx
