+++ src/main.jsx
@@ -1,0 +1,28 @@
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import { createBrowserRouter, RouterProvider } from 'react-router-dom'
+
+import './index.css'
+
+import App from './App.jsx'
+import ErrorPage from './pages/Error'
+import Homepage from './pages/Homepage/index.jsx'
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <App />,
+    errorElement: <ErrorPage />,
+    children: [
+      {
+        index: true,
+        element: <Homepage />
+      }
+    ]
+  }
+])
+
+
+ReactDOM.createRoot(document.getElementById('root')).render(
+  <RouterProvider router= {router} />
+)
